[220213 - LHG] - Bookmark 클래스 추가 및 Member와 연관관계 매핑 - Ban, Member 관계 수정 - StatusType 추가 (Member, Board, Post, Comment) - PostHashtag 패키지 경로 수정 (hashtag -> post) - PostLikey 패키지 경로 수정 (likey -> post) - CommentLikey 패키지 경로 수정 (likey -> comment)
