+++ src/main/java/com/ootd/with/domain/board/Board.java
@@ -1,5 +1,6 @@
 package com.ootd.with.domain.board;
 
+import com.ootd.with.domain.enumtype.StatusType;
 import com.ootd.with.domain.post.Post;
 import lombok.Builder;
 import lombok.Getter;
@@ -23,6 +24,9 @@ public class Board {
 
     @OneToMany(mappedBy = "board")
     private List<Post> posts = new ArrayList<>();
+
+    @Enumerated(EnumType.STRING)
+    private StatusType statusType;
 
     @Builder
     public Board(String name) {
