+++ src/main/java/com/ootd/with/domain/enumtype/StatusType.java
@@ -1,0 +1,7 @@
+package com.ootd.with.domain.enumtype;
+
+public enum StatusType {
+    NORMAL,   // 정상 상태
+    HIDDEN,   // 숨김 상태 (노출되지 않음)
+    DELETED   // 삭제 상태 (삭제된 데이터도 보관, 관리하기 위함)
+}
