+++ src/main/java/com/ootd/with/domain/post/PostHashtagRepository.java
@@ -1,4 +1,4 @@
-package com.ootd.with.domain.hashtag;
+package com.ootd.with.domain.post;
 
 import org.springframework.data.jpa.repository.JpaRepository;
